Break out the `logs` command
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`logs` command into its own subpackage.
