+++ pkg/cmd/pulumi/logs/logs.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package logs
 
 import (
 	"fmt"
@@ -38,7 +38,7 @@ import (
 	"github.com/pulumi/pulumi/sdk/v3/go/common/workspace"
 )
 
-func newLogsCmd() *cobra.Command {
+func NewLogsCmd() *cobra.Command {
 	var stackName string
 	var follow bool
 	var since string
